Replace hill-climb blend optimizer with Caruana greedy ensemble selection
Implements iterative greedy candidate selection (with replacement) using
score_predictions from cv.py with the configured primary_metric, so the
optimizer works correctly for regression and binary classification.

Key changes:
- New script: scripts/select_greedy_ensemble.py
- Deleted: scripts/optimize_blend_weights.py
- Uses fast_score closure to precompute y_true_binary once, avoiding repeated
  pd.Series/pd.unique overhead in the inner loop
- Incremental blend computation (current_sum + oof_matrix[i]) instead of
  full np.average per candidate
- tqdm bars on OOF loading and greedy selection rounds
- Early stopping when no candidate improves (avoids unnecessary rounds)
- --n-rounds arg (default 100); OOF disk cache and config snippet unchanged

Closes #225
